+++ lib/src/buffer_image.dart
@@ -1,6 +1,5 @@
 library buffer_image;
 
-import 'dart:io';
 import 'dart:math';
 import 'dart:typed_data';
 import 'dart:ui';
@@ -38,9 +37,9 @@ class BufferImage {
         image.height);
   }
 
-  /// load image from a `file` use system codec([decodeImageFromList])
-  static Future<BufferImage?> fromFile(File file) async {
-    return fromImage(await decodeImageFromList(file.readAsBytesSync()));
+  /// load image from a image `fileData` use system codec([decodeImageFromList])
+  static Future<BufferImage?> fromFile(Uint8List fileData) async {
+    return fromImage(await decodeImageFromList(fileData));
   }
 
   int get width {
@@ -51,22 +50,50 @@ class BufferImage {
     return _height;
   }
 
-  /// set the [Color] at Offset(`x`, `y`)
+  /// set the [Color] at Offset([x], [y])
   setColor(int x, int y, Color color) {
+    assert(x >= 0 && x < width, 'x($x) out of with boundary(0 - $width)');
+    assert(y >= 0 && y < height, 'y($y) out of height boundary(0 - $height)');
     _buffer[y * _width * bytePerPixel + x * bytePerPixel] = color.red;
     _buffer[y * _width * bytePerPixel + x * bytePerPixel + 1] = color.green;
     _buffer[y * _width * bytePerPixel + x * bytePerPixel + 2] = color.blue;
     _buffer[y * _width * bytePerPixel + x * bytePerPixel + 3] = color.alpha;
   }
 
-  /// get the [Color] at Offset(`x`, `y`)
+  /// set [Color] at Offset([x], [y]) without error
+  setColorSafe(int x, int y, Color color){
+    if(x >= 0 && x < width && y >= 0 && y < height){
+      setColor(x, y, color);
+    }
+  }
+
+  /// get the [Color] at Offset([x], [y])
   Color getColor(int x, int y) {
+    assert(x >= 0 && x < width, 'x($x) out of with boundary(0 - $width)');
+    assert(y >= 0 && y < height, 'y($y) out of height boundary(0 - $height)');
     return Color.fromARGB(
       _buffer[y * _width * bytePerPixel + x * bytePerPixel + 3],
       _buffer[y * _width * bytePerPixel + x * bytePerPixel],
       _buffer[y * _width * bytePerPixel + x * bytePerPixel + 1],
       _buffer[y * _width * bytePerPixel + x * bytePerPixel + 2],
     );
+  }
+
+  /// get the [Color] at Offset([x], [y]) with out error
+  ///
+  /// if out of boundary return [defaultColor]
+  /// if defaultColor is `null` return nearest boundary color
+  Color getColorSafe(int x, int y, [Color? defaultColor = const Color(0x00ffffff)]){
+    if(x >= 0 && x < width && y >= 0 && y < height){
+      return getColor(x, y);
+    }else if(defaultColor == null){
+      if(x < 0) x = 0;
+      if(x > width - 1) x = width - 1;
+      if(y < 0) y = 0;
+      if(y > height - 1) y = height - 1;
+      return getColor(x, y);
+    }
+    return defaultColor;
   }
 
   /// scale by `ratio` width `sample`
@@ -112,9 +139,10 @@ class BufferImage {
   /// Else adjust the canvas to fit the rotated image
   /// `isAntialias` not implemented
   rotate(double radian,
-      [bool isAntialias = true,
+  {bool isAntialias = true,
+      SampleMode sample = SampleMode.bilinear,
       Color bgColor = const Color.fromARGB(0, 255, 255, 255),
-      bool isClip = false]) {
+      bool isClip = false}) {
     int newWidth = _width;
     int newHeight = _height;
     if (!isClip) {
@@ -127,12 +155,11 @@ class BufferImage {
     double nxr = (newWidth - 1) / 2;
     double nyr = (newHeight - 1) / 2;
 
-    //List<String> logs = [];
     //rotate
     for (int x = 0; x < newWidth; x++) {
       for (int y = 0; y < newHeight; y++) {
-        int xPos = xr.round();
-        int yPos = yr.round();
+        double xPos = xr;
+        double yPos = yr;
 
         // 非中心点才可以计算
         if (x != nxr || y != nyr) {
@@ -149,14 +176,15 @@ class BufferImage {
           }
           double newRadian = curRadian - radian;
 
-          xPos = (cos(newRadian) * r + xr).round();
-          yPos = (_height - sin(newRadian) * r - yr).round();
+          xPos = (cos(newRadian) * r + xr);
+          yPos = (_height - sin(newRadian) * r - yr);
         }
-        Point<int> orig = Point(xPos, yPos);
+        Point<double> orig = Point(xPos, yPos);
         Color newColor = bgColor;
-        if (orig.x >= 0 && orig.x < _width && orig.y >= 0 && orig.y < _height) {
+        if (orig.x > -1 && orig.x < _width && orig.y > -1 && orig.y < _height) {
           //logs.add("$x, $y => ${orig.x}, ${orig.y}");
-          newColor = getColor(orig.x, orig.y);
+          //newColor = getColor(orig.x, orig.y);
+          newColor = sample.sample(orig, this, bgColor);
         }
         newBuffer[y * newWidth * bytePerPixel + x * bytePerPixel] =
             newColor.red;
@@ -224,6 +252,34 @@ class BufferImage {
           setColor(
               x, y, blend.blend(image.getColor(oPoint.x, oPoint.y), oColor));
         }
+      }
+    }
+  }
+
+  drawRect(Rect rect, Color color, [BlendMode mode = BlendMode.src]){
+    BlendModeAction blend = BlendModeAction(mode);
+    int maxX = min(_width, rect.right.round());
+    int minY = max(0, rect.top.round());
+    int maxY = min(_height, rect.bottom.round());
+    for (int x = max(0, rect.left.round()); x < maxX; x++) {
+      for (int y = minY; y < maxY; y++) {
+        Color oColor = getColor(x, y);
+        setColor(x, y, blend.blend(color, oColor));
+      }
+    }
+  }
+
+  drawImage(BufferImage image, Offset offset, [BlendMode mode = BlendMode.src]){
+    BlendModeAction blend = BlendModeAction(mode);
+    int minX = max(0, offset.dx.round());
+    int maxX = min(_width, (offset.dx + image.width).round());
+    int minY = max(0, offset.dy.round());
+    int maxY = min(_height, (offset.dy + image.height).round());
+
+    for (int x = minX; x < maxX; x++) {
+      for (int y = minY; y < maxY; y++) {
+        Color oColor = getColor(x, y);
+        setColor(x, y, blend.blend(image.getColor(x - minX, y - minY), oColor));
       }
     }
   }