+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+## 0.1.0
+
+* add drawRect, drawImage
+* add sample mode for rotate
+* add getColorSafe & setColorSafe
+
 ## 0.0.2
+
 * comments & package
 
 ## 0.0.1